agent/requirements.txt
livekit-agents[openai]
livekit-plugins-simli
elevenlabs
python-dotenv